+++ base/AirModel.ts
@@ -4,10 +4,10 @@ import { AirFormFieldConfig } from '../config/AirFormFieldConfig'
 import { AirSearchFieldConfig } from '../config/AirSearchFieldConfig'
 import { AirTableFieldConfig } from '../config/AirTableFieldConfig'
 import { getModelConfig } from '../decorator/Model'
-import { getFormConfig, getFormConfigList } from '../decorator/FormField'
+import { getFormConfig, getFormConfigList } from '../decorator/Form'
 import { getFieldConfig, getToJson, getToModel } from '../decorator/Field'
-import { getSearchConfigList } from '../decorator/SearchField'
-import { getTableConfigList } from '../decorator/TableField'
+import { getSearchConfigList } from '../decorator/Search'
+import { getTableConfigList } from '../decorator/Table'
 import { IFieldConfig } from '../interface/decorators/IFieldConfig'
 import { IJson } from '../interface/IJson'
 import { IModelConfig } from '../interface/decorators/IModelConfig'
